add usePostDetail hook, add loading post form
